salesforce-appflow/iac/requirements.txt
attrs==20.3.0
aws-cdk.alexa-ask==1.98.0
aws-cdk.assets==1.98.0
aws-cdk.aws-accessanalyzer==1.98.0
aws-cdk.aws-acmpca==1.98.0
aws-cdk.aws-amazonmq==1.98.0
aws-cdk.aws-amplify==1.98.0
aws-cdk.aws-apigateway==1.98.0
aws-cdk.aws-apigatewayv2==1.98.0
aws-cdk.aws-appconfig==1.98.0
aws-cdk.aws-appflow==1.98.0
aws-cdk.aws-applicationautoscaling==1.98.0
aws-cdk.aws-applicationinsights==1.98.0
aws-cdk.aws-appmesh==1.98.0
aws-cdk.aws-appstream==1.98.0
aws-cdk.aws-appsync==1.98.0
aws-cdk.aws-athena==1.98.0
aws-cdk.aws-auditmanager==1.98.0
aws-cdk.aws-autoscaling==1.98.0
aws-cdk.aws-autoscaling-common==1.98.0
aws-cdk.aws-autoscaling-hooktargets==1.98.0
aws-cdk.aws-autoscalingplans==1.98.0
aws-cdk.aws-backup==1.98.0
aws-cdk.aws-batch==1.98.0
aws-cdk.aws-budgets==1.98.0
aws-cdk.aws-cassandra==1.98.0
aws-cdk.aws-ce==1.98.0
aws-cdk.aws-certificatemanager==1.98.0
aws-cdk.aws-chatbot==1.98.0
aws-cdk.aws-cloud9==1.98.0
aws-cdk.aws-cloudformation==1.98.0
aws-cdk.aws-cloudfront==1.98.0
aws-cdk.aws-cloudtrail==1.98.0
aws-cdk.aws-cloudwatch==1.98.0
aws-cdk.aws-cloudwatch-ACTIONS==1.98.0
aws-cdk.aws-codeartifact==1.98.0
aws-cdk.aws-codebuild==1.98.0
aws-cdk.aws-codecommit==1.98.0
aws-cdk.aws-codedeploy==1.98.0
aws-cdk.aws-codeguruprofiler==1.98.0
aws-cdk.aws-codegurureviewer==1.98.0
aws-cdk.aws-codepipeline==1.98.0
aws-cdk.aws-codestar==1.98.0
aws-cdk.aws-codestarconnections==1.98.0
aws-cdk.aws-codestarnotifications==1.98.0
aws-cdk.aws-cognito==1.98.0
aws-cdk.aws-config==1.98.0
aws-cdk.aws-databrew==1.98.0
aws-cdk.aws-datapipeline==1.98.0
aws-cdk.aws-datasync==1.98.0
aws-cdk.aws-dax==1.98.0
aws-cdk.aws-detective==1.98.0
aws-cdk.aws-devopsguru==1.98.0
aws-cdk.aws-directoryservice==1.98.0
aws-cdk.aws-dlm==1.98.0
aws-cdk.aws-dms==1.98.0
aws-cdk.aws-docdb==1.98.0
aws-cdk.aws-dynamodb==1.98.0
aws-cdk.aws-ec2==1.98.0
aws-cdk.aws-ecr==1.98.0
aws-cdk.aws-ecr-assets==1.98.0
aws-cdk.aws-ecs==1.98.0
aws-cdk.aws-efs==1.98.0
aws-cdk.aws-eks==1.98.0
aws-cdk.aws-elasticache==1.98.0
aws-cdk.aws-elasticbeanstalk==1.98.0
aws-cdk.aws-elasticloadbalancing==1.98.0
aws-cdk.aws-elasticloadbalancingv2==1.98.0
aws-cdk.aws-elasticsearch==1.98.0
aws-cdk.aws-emr==1.98.0
aws-cdk.aws-emrcontainers==1.98.0
aws-cdk.aws-events==1.98.0
aws-cdk.aws-events-targets==1.98.0
aws-cdk.aws-eventschemas==1.98.0
aws-cdk.aws-glue==1.98.0
aws-cdk.aws-guardduty==1.98.0
aws-cdk.aws-iam==1.98.0
aws-cdk.aws-imagebuilder==1.98.0
aws-cdk.aws-inspector==1.98.0
aws-cdk.aws-ivs==1.98.0
aws-cdk.aws-kinesis==1.98.0
aws-cdk.aws-kinesisanalytics==1.98.0
aws-cdk.aws-kinesisfirehose==1.98.0
aws-cdk.aws-kms==1.98.0
aws-cdk.aws-lakeformation==1.98.0
aws-cdk.aws-lambda==1.98.0
aws-cdk.aws-lambda-event-sources==1.98.0
aws-cdk.aws-licensemanager==1.98.0
aws-cdk.aws-logs==1.98.0
aws-cdk.aws-pinpoint==1.98.0
aws-cdk.aws-pinpointemail==1.98.0
aws-cdk.pipelines==1.98.0
aws-cdk.aws-qldb==1.98.0
aws-cdk.aws-quicksight==1.98.0
aws-cdk.aws-ram==1.98.0
aws-cdk.aws-rds==1.98.0
aws-cdk.aws-redshift==1.98.0
aws-cdk.aws-resourcegroups==1.98.0
aws-cdk.aws-robomaker==1.98.0
aws-cdk.aws-route53==1.98.0
aws-cdk.aws-route53-targets==1.98.0
aws-cdk.aws-route53resolver==1.98.0
aws-cdk.aws-s3==1.98.0
aws-cdk.aws-s3-deployment==1.98.0
aws-cdk.aws-s3-assets==1.98.0
aws-cdk.aws-s3objectlambda==1.98.0
aws-cdk.aws-s3outposts==1.98.0
aws-cdk.aws-sagemaker==1.98.0
aws-cdk.aws-sam==1.98.0
aws-cdk.aws-sdb==1.98.0
aws-cdk.aws-secretsmanager==1.98.0
aws-cdk.aws-securityhub==1.98.0
aws-cdk.aws-servicecatalog==1.98.0
aws-cdk.aws-servicecatalogappregistry==1.98.0
aws-cdk.aws-servicediscovery==1.98.0
aws-cdk.aws-ses==1.98.0
aws-cdk.aws-signer==1.98.0
aws-cdk.aws-sns==1.98.0
aws-cdk.aws-sns-subscriptions==1.98.0
aws-cdk.aws-sqs==1.98.0
aws-cdk.aws-ssm==1.98.0
aws-cdk.aws-sso==1.98.0
aws-cdk.aws-stepfunctions==1.98.0
aws-cdk.aws-timestream==1.98.0
aws-cdk.aws-transfer==1.98.0
aws-cdk.cloud-assembly-schema==1.98.0
aws-cdk.cloudformation-include==1.98.0
aws-cdk.core==1.98.0
aws-cdk.custom-resources==1.98.0
aws-cdk.cx-api==1.98.0
aws-cdk.lambda-layer-awscli==1.98.0
aws-cdk.region-info==1.98.0
cattrs==1.5.0
constructs==3.3.77
jsii==1.30.0
publication==0.0.3
python-dateutil==2.8.1
six==1.16.0
typing-extensions==3.10.0.0
pydantic
PyYAML
jinja2
aws-lambda-powertools